chore: fixed the tests + references and deleted the olds
