+++ packages/salesforcedx-vscode-apex-debugger/test/vscode-integration/adapter/MockApexExtension.ts
@@ -22,6 +22,14 @@ export class MockApexExtension implements Extension<any> {
     return Promise.resolve('activated');
   }
   public exports: any;
+
+  public languageClientManager = {
+    getStatus: () => ({
+      isReady: () => true,
+      failedToInitialize: () => false,
+      getStatusMessage: () => ''
+    })
+  };
 }
 
 class MockJorje {
@@ -36,12 +44,4 @@ class MockJorje {
     ];
     return Promise.resolve(response);
   }
-
-  public languageClientUtils = {
-    getStatus: () => ({
-      isReady: () => true,
-      failedToInitialize: () => false,
-      getStatusMessage: () => ''
-    })
-  };
 }
