+++ packages/salesforcedx-vscode-apex/src/commands/metadataOrchestrator.ts
@@ -7,7 +7,7 @@
 import { LLMServiceInterface, ServiceProvider, ServiceType } from '@salesforce/vscode-service-provider';
 import * as path from 'path';
 import * as vscode from 'vscode';
-import { languageClientUtils } from '../languageUtils';
+import { languageClientManager } from '../languageUtils';
 import { nls } from '../messages';
 import GenerationInteractionLogger from '../oas/generationInteractionLogger';
 import {
@@ -64,7 +64,7 @@ export class MetadataOrchestrator {
     gil.addApexClassOASEligibleRequest(requests.payload);
     const telemetryService = await getTelemetryService();
     let response;
-    const languageClient = languageClientUtils.getClientInstance();
+    const languageClient = languageClientManager.getClientInstance();
     if (languageClient) {
       const classNumbers = requests.payload.length.toString();
       const requestTarget = this.requestTarget(requests);
@@ -91,7 +91,7 @@ export class MetadataOrchestrator {
   ): Promise<ApexClassOASGatherContextResponse | undefined> => {
     const telemetryService = await getTelemetryService();
     let response: ApexClassOASGatherContextResponse | undefined;
-    const languageClient = languageClientUtils.getClientInstance();
+    const languageClient = languageClientManager.getClientInstance();
     if (languageClient) {
       try {
         response = await languageClient?.gatherOpenAPIContext(
