+++ packages/salesforcedx-vscode-apex/src/index.ts
@@ -38,9 +38,9 @@ import {
   getApexTests,
   getExceptionBreakpointInfo,
   getLineBreakpointInfo,
-  languageClientUtils
+  languageClientManager
 } from './languageUtils';
-import { restartLanguageServerAndClient, createLanguageClient } from './languageUtils/languageClientUtils';
+import { restartLanguageServerAndClient, createLanguageClient } from './languageUtils';
 import { nls } from './messages';
 import { checkIfESRIsDecomposed } from './oasUtils';
 import { getTelemetryService } from './telemetry/telemetry';
@@ -84,7 +84,7 @@ export const activate = async (context: vscode.ExtensionContext) => {
 
   // start the language server and client
   const languageServerStatusBarItem = new ApexLSPStatusBarItem();
-  languageClientUtils.setStatusBarInstance(languageServerStatusBarItem);
+  languageClientManager.setStatusBarInstance(languageServerStatusBarItem);
   await createLanguageClient(extensionContext, languageServerStatusBarItem);
 
   // Javadoc support
@@ -115,7 +115,7 @@ export const activate = async (context: vscode.ExtensionContext) => {
     getLineBreakpointInfo,
     getExceptionBreakpointInfo,
     getApexTests,
-    languageClientUtils
+    languageClientManager
   };
 
   void activationTracker.markActivationStop(new Date());
@@ -266,7 +266,7 @@ const registerTestView = (): vscode.Disposable => {
   // Refresh Test View command
   testViewItems.push(
     vscode.commands.registerCommand(`${testOutlineProvider.getId()}.refresh`, () => {
-      if (languageClientUtils.getStatus().isReady()) {
+      if (languageClientManager.getStatus().isReady()) {
         return testOutlineProvider.refresh();
       }
     })
@@ -282,7 +282,7 @@ const registerTestView = (): vscode.Disposable => {
 };
 
 export const deactivate = async () => {
-  await languageClientUtils.getClientInstance()?.stop();
+  await languageClientManager.getClientInstance()?.stop();
   const telemetryService = await getTelemetryService();
   telemetryService?.sendExtensionDeactivationEvent();
 };
