+++ packages/salesforcedx-vscode-apex/src/languageUtils/index.ts
@@ -4,24 +4,53 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
+import * as vscode from 'vscode';
+import { ApexLanguageClient } from '../apexLanguageClient';
+import ApexLSPStatusBarItem from '../apexLspStatusBarItem';
 import { LanguageClientManager, ClientStatus, LanguageClientStatus, ProcessDetail } from './languageClientManager';
-import { LanguageClientUtils } from './languageClientUtils';
 
 export const languageClientManager = LanguageClientManager.getInstance();
-export const languageClientUtils = LanguageClientUtils.getInstance();
+
 export { ClientStatus, LanguageClientStatus, ProcessDetail };
 
-export {
-  getLineBreakpointInfo,
-  getApexTests,
-  getExceptionBreakpointInfo,
-  restartLanguageServerAndClient,
-  createLanguageClient,
-  indexerDoneHandler
-} from './languageClientUtils';
+export const getLineBreakpointInfo = async (): Promise<{}> =>
+  LanguageClientManager.getInstance().getLineBreakpointInfo();
+
+export const getApexTests = async (): Promise<any[]> => LanguageClientManager.getInstance().getApexTests();
+
+export const getExceptionBreakpointInfo = async (): Promise<{}> =>
+  LanguageClientManager.getInstance().getExceptionBreakpointInfo();
+
+export const restartLanguageServerAndClient = async (extensionContext: vscode.ExtensionContext): Promise<void> =>
+  LanguageClientManager.getInstance().restartLanguageServerAndClient(extensionContext);
+
+export const createLanguageClient = async (
+  extensionContext: vscode.ExtensionContext,
+  languageServerStatusBarItem: ApexLSPStatusBarItem
+): Promise<void> =>
+  LanguageClientManager.getInstance().createLanguageClient(extensionContext, languageServerStatusBarItem);
+
+export const indexerDoneHandler = async (
+  enableSyncInitJobs: boolean,
+  languageClient: ApexLanguageClient,
+  languageServerStatusBarItem: ApexLSPStatusBarItem
+): Promise<void> =>
+  LanguageClientManager.getInstance().indexerDoneHandler(
+    enableSyncInitJobs,
+    languageClient,
+    languageServerStatusBarItem
+  );
+
+export const findAndCheckOrphanedProcesses = async (): Promise<ProcessDetail[]> =>
+  LanguageClientManager.getInstance().findAndCheckOrphanedProcesses();
+
+export const terminateProcess = (pid: number): void => {
+  LanguageClientManager.getInstance().terminateProcess(pid);
+};
+
+export const canRunCheck = async (isWindows: boolean): Promise<boolean> =>
+  LanguageClientManager.getInstance().canRunCheck(isWindows);
 
 export { configureApexLanguage } from './apexLanguageConfiguration';
-
-export { languageServerUtils } from './languageServerUtils';
 
 export { extensionUtils } from './extensionUtils';
