+++ packages/salesforcedx-vscode-apex/test/jest/languageUtils/extensionUtils.test.ts
@@ -5,7 +5,7 @@
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import ApexLSPStatusBarItem from '../../../src/apexLspStatusBarItem';
-import { ClientStatus, languageClientUtils } from '../../../src/languageUtils';
+import { ClientStatus, languageClientManager } from '../../../src/languageUtils';
 import { extensionUtils } from '../../../src/languageUtils';
 import * as testOutlineProvider from '../../../src/views/testOutlineProvider';
 
@@ -32,7 +32,7 @@ describe('extensionUtils Unit Tests.', () => {
     languageServerStatusBarItem = new ApexLSPStatusBarItem();
     refreshSpy = jest.spyOn(mockTestOutlineProviderInst, 'refresh').mockResolvedValue();
     readySpy = jest.spyOn(languageServerStatusBarItem, 'ready');
-    setStatusSpy = jest.spyOn(languageClientUtils, 'setStatus').mockReturnValue();
+    setStatusSpy = jest.spyOn(languageClientManager, 'setStatus').mockReturnValue();
     mockLanguageClient = {
       errorHandler: {
         serviceHasStartedSuccessfully: jest.fn()
